Change tree on the left to logical representation of the project
As opposed to physical representation.

Physically, the directory structure must be the following:

o Contenuti <- is a directory. Contains the content-page subdirs
├o 1 <- is a directory, can have any name
│  ├o content.md <- required, is the actual content of the page
│  ├o page.json <- page descriptor file.
│  │  Contains title, header_image and shared_text
│  ├o header.png <- required, header image. Can have any name
│  ├o img_1.png <- optional, image file, can have any name
│  ┊
│  └o img_n.png <- every page can have any number of images.
┊     They are references inside content.md
├o n <- can have any number of pages in a project
o Glossary <- is a directory. Contains the glossary-page subdirs

The structure of the Glossary directory is yet to be defined, although
it will be similar to that of Contenuti