First iteration
Copied files have been changed to have an uuid as name.

Kown bugs:
* pdfFile inside page.json refers to the file with the old name
    instead of uuid
* headerImage inside page.json refers to the file with the old name
    instead of uuid
* image links inside saved markdown are relative to the binary, they
    must be relative to the page directory (so, just their uuid name)

The last bug implies that, when fixed, when the markdown is loaded the
img src attribute must be modified replacing it with the absolute full
path on-the-fly (since it will be different across computers etc), which
implies that when saving all the img src attributes will contain
absolute paths, so the current logic of "if this path is relative it's
inside the page dir, external otherwise" will need to be changed to
"if the page dir path is a prefix of the full absolute path of the
image, that image is internal to the page dir, external otherwise".